requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
PyYAML==6.0